добавлена секция egts+

+++ errors.go
@@ -1,118 +1,118 @@
 package main
 
 //код сообщения, что пакет успешно обработано
-const egtsPcOk = 0
+const egtsPcOk = uint8(0)
 
 //код сообщения, что пакет в процессе обработки (результат обработки ещё не известен)
-const egtsPcInProgress = 1
+const egtsPcInProgress = uint8(1)
 
 //неподдерживаемый протокол
-const egtsPcUnsProtocol = 128
+const egtsPcUnsProtocol = uint8(128)
 
 //ошибка декодирования
-const egtsPcDecryptError = 129
+const egtsPcDecryptError = uint8(129)
 
 //обработка запрещена
-const egtsPcProcDenied = 130
+const egtsPcProcDenied = uint8(130)
 
 //неверный формат заголовка
-const egtsPcIncHeaderform = 131
+const egtsPcIncHeaderform = uint8(131)
 
 //неверный формат данных
-const egtsPcIncDataform = 132
+const egtsPcIncDataform = uint8(132)
 
 //EgtsPcUnsType неподдерживаемый тип
-const egtsPcUnsType = 133
+const egtsPcUnsType = uint8(133)
 
 //неверное количество параметров
-const egtsPcNotenParams = 134
+const egtsPcNotenParams = uint8(134)
 
 //попытка повторной обработки
-const egtsPcDblProc = 135
+const egtsPcDblProc = uint8(135)
 
 //обработка данных от источника запрещена
-const egtsPcProcSrcDenied = 136
+const egtsPcProcSrcDenied = uint8(136)
 
 //ошибка контрольной суммы заголовка
-const egtsPcHeadercrcError = 137
+const egtsPcHeadercrcError = uint8(137)
 
 //ошибка контрольной суммы данных
-const egtsPcDatacrcError = 138
+const egtsPcDatacrcError = uint8(138)
 
 //некорректная длина данных
-const egtsPcInvdatalen = 139
+const egtsPcInvdatalen = uint8(139)
 
 //маршрут не найден
-const egtsPcRouteNfound = 140
+const egtsPcRouteNfound = uint8(140)
 
 //маршрут закрыт
-const egtsPcRouteClosed = 141
+const egtsPcRouteClosed = uint8(141)
 
 //маршрутизация запрещена
-const egtsPcRouteDenied = 142
+const egtsPcRouteDenied = uint8(142)
 
 //неверный адрес
-const egtsPcInvaddr = 143
+const egtsPcInvaddr = uint8(143)
 
 //превышено количество ретрансляции данных
-const egtsPcTtlexpired = 144
+const egtsPcTtlexpired = uint8(144)
 
 //нет подтверждения
-const egtsPcNoAck = 145
+const egtsPcNoAck = uint8(145)
 
 //объект не найден
-const egtsPcObjNfound = 146
+const egtsPcObjNfound = uint8(146)
 
 //событие не найдено
-const egtsPcEvntNfound = 147
+const egtsPcEvntNfound = uint8(147)
 
 //сервис не найден
-const egtsPcSrvcNfound = 148
+const egtsPcSrvcNfound = uint8(148)
 
 //сервис запрещён
-const egtsPcSrvcDenied = 149
+const egtsPcSrvcDenied = uint8(149)
 
 //неизвестный тип сервиса
-const egtsPcSrvcUnkn = 150
+const egtsPcSrvcUnkn = uint8(150)
 
 //авторизация запрещена
-const egtsPcAuthPenied = 151
+const egtsPcAuthPenied = uint8(151)
 
 //объект уже существует
-const egtsPcAlreadyExists = 152
+const egtsPcAlreadyExists = uint8(152)
 
 //идентификатор не найден
-const egtsPcIDNfound = 153
+const egtsPcIDNfound = uint8(153)
 
 //неправильная дата и время
-const egtsPcIncDatetime = 154
+const egtsPcIncDatetime = uint8(154)
 
 //ошибка ввода/вывода
-const egtsPcIoError = 155
+const egtsPcIoError = uint8(155)
 
 //недостаточно ресурсов
-const egtsPcNoResAvail = 156
+const egtsPcNoResAvail = uint8(156)
 
 //внутренний сбой модуля
-const egtsPcModuleFault = 157
+const egtsPcModuleFault = uint8(157)
 
 //сбой в работе цепи питания модуля
-const egtsPcModulePwrFlt = 158
+const egtsPcModulePwrFlt = uint8(158)
 
 //сбой в работе микроконтроллера модуля
-const egtsPcModuleProcFlt = 159
+const egtsPcModuleProcFlt = uint8(159)
 
 //сбой в работе программы модуля
-const egtsPcModuleSwFlt = 160
+const egtsPcModuleSwFlt = uint8(160)
 
 //сбой в работе внутреннего ПО модуля
-const egtsPcModuleFwFlt = 161
+const egtsPcModuleFwFlt = uint8(161)
 
 //сбой в работе блока ввода/вывода модуля
-const egtsPcModuleIoFlt = 162
+const egtsPcModuleIoFlt = uint8(162)
 
 //сбой в работе внутренней памяти модуля
-const egtsPcModuleMemFlt = 163
+const egtsPcModuleMemFlt = uint8(163)
 
 //тест не пройден
-const egtsPcTestFailed = 164
+const egtsPcTestFailed = uint8(164)
